Ensure that exact matches takes precedence in CheckGuess
Found a bug in CheckGuess where a code like "3 3 0 3" and a guess like "3 3 3 3" would
result in 3 exact hits + 1 value hit - which is incorrect.

The 3rd "3" in the guess would count as a value hit, because at that point in the analysis, the count for "3" would be 1 because the next exact match had not been seen yet.